feat(settings): Improve settings UI and add new tabs
This commit introduces several improvements to the settings UI in the desktop
application. The main changes are:

1. Added new tabs for Profile, AI, Calendar, Notifications, Templates,
   Extensions, Team, and Billing.
2. Implemented a dynamic tab title generation function to handle the new tabs.
3. Removed the `SettingsHeader` component as its functionality has been
   integrated into the main `Component` function.
4. Removed the unused `MainSidebar` component.
5. Utilized the `useLingui` hook to provide localized tab titles.

These changes aim to enhance the user experience by providing a more
comprehensive and organized settings interface, allowing users to easily
navigate and manage various aspects of the application.
