+++ apps/desktop/src/components/settings/components/index.ts
@@ -1,5 +1,3 @@
-export * from "./main-sidebar";
-export * from "./settings-header";
 export * from "./tab-icon";
 export * from "./templates-sidebar";
 export * from "./types";
